Add some automated TestFlight abilities to the build logic.
This is a first pass at adding some automated TestFlight stuff to the build logic.

This functionality might not actually work as of this commit, and I haven't tested it because I don't have the Syntellia TestFlight key, which is required to automatically push the built IPA to TestFlight.  However, these scripts and logic were taken from an older iOS app where all this TestFlight machinery worked.

This commit also adds the script foo needed to add git shorthash tags to the apps version variable.