Fix for `FL_BUILD_NUMBER` not getting expanded / set in `Info.plist`.
This (hopefully) fixes the problem where `FL_BUILD_NUMBER` was not getting expanded / set in `Info.plist`.
This was done via a kludgey hack- the issue seems to be related to a "bug" in the way Xcode evaluates and incorporates results from `.xconfig` files.
The hack was to change the name of the variable that gets set in a shell script, then setting `FL_BUILD_NUMBER` to `${FL_BUILDINFO_BUILD_NUMBER}` at the TOP LEVEL project settings, which seems to be sequenced and evaluated before the build targets settings.
This "seems" to make the value of the variable be set before the `Info.plist` expansion takes place.

Kludgey.