Fixed text always being hidden behind input view.
With iOS 8, the inputView window always has a blur effect hiding what’s
behind it.
Since we use a fullscreen inputView, the workaround is to make the
inputView have an alpha mask that is 1 towards the bottom (keyboard),
and 0 towards the top. Note that this hides the action button for
non-VoiceOver users.